misc/qt-updater/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --all-extras --output-file=requirements.txt pyproject.toml
#
black==23.9.1
    # via qt-updater (pyproject.toml)
build==1.0.3
    # via pip-tools
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.0
    # via requests
click==8.1.7
    # via
    #   black
    #   pip-tools
idna==3.4
    # via requests
mypy==1.6.0
    # via qt-updater (pyproject.toml)
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
packaging==23.2
    # via
    #   black
    #   build
pathspec==0.11.2
    # via black
pip-tools==7.3.0
    # via qt-updater (pyproject.toml)
platformdirs==3.11.0
    # via black
pyproject-hooks==1.0.0
    # via build
requests==2.31.0
    # via qt-updater (pyproject.toml)
ruamel-yaml==0.17.35
    # via qt-updater (pyproject.toml)
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
ruff==0.0.292
    # via qt-updater (pyproject.toml)
types-requests==2.31.0.8
    # via qt-updater (pyproject.toml)
typing-extensions==4.8.0
    # via mypy
urllib3==2.0.6
    # via
    #   requests
    #   types-requests
wheel==0.41.2
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
